+++ src/components/canvas/PotoCanvasClip.css
@@ -1,11 +1,11 @@
-.PotoCanvasClipper {
+.PotoCanvasClip {
   --shadow-color: #3339;
 
   left: var(--app-space);
   position: absolute;
   top: var(--app-space);
 }
-  .PotoCanvasClipper-shadow {
+  .PotoCanvasClip-shadow {
     background-color: var(--shadow-color);
     position: absolute;
   }
